initial commit with JWT logic in controller

+++ src/users_crud/controller.ts
@@ -6,10 +6,8 @@ import { TYPES } from "../types";
 import { IUserCrudController } from "./controller.interface";
 import "dotenv/config";
 import { IUserCrudService } from "./service.interface";
-import { CheckAccessToken } from "../common/middlewares/checkJWT.middleware";
-import { verify } from "jsonwebtoken";
-import { IUserPayload } from "../users_auth/service.interface";
 import { CheckIfAdmin } from "../common/middlewares/checkIfAdmin.middleware";
+import { IJWTService } from "../users_auth/jwt.service.interface";
 
 @injectable()
 export class UserCrudController
@@ -17,7 +15,8 @@ export class UserCrudController
 	implements IUserCrudController
 {
 	constructor(
-		@inject(TYPES.UserCrudService) private userService: IUserCrudService
+		@inject(TYPES.UserCrudService) private userService: IUserCrudService,
+		@inject(TYPES.JWTService) private jwtService: IJWTService
 	) {
 		super();
 		this.bindRoutes([
@@ -25,25 +24,35 @@ export class UserCrudController
 				path: "/getAll",
 				method: "get",
 				func: this.getAllForAdmin,
-				middlewares: [new CheckAccessToken(), new CheckIfAdmin()],
+				middlewares: [new CheckIfAdmin()],
 			},
 			{
 				path: "/getOne/:id",
 				method: "get",
 				func: this.getOneForAdmin,
-				middlewares: [new CheckAccessToken(), new CheckIfAdmin()],
+				middlewares: [new CheckIfAdmin()],
 			},
 			{
 				path: "/update",
 				method: "patch",
 				func: this.update,
-				middlewares: [new CheckAccessToken()],
+			},
+			{
+				path: "/updateAny",
+				method: "patch",
+				func: this.updateAny,
+				middlewares: [new CheckIfAdmin()],
 			},
 			{
 				path: "/delete",
 				method: "delete",
 				func: this.delete,
-				middlewares: [new CheckAccessToken()],
+			},
+			{
+				path: "/deleteAny",
+				method: "delete",
+				func: this.deleteAny,
+				middlewares: [new CheckIfAdmin()],
 			},
 		]);
 	}
@@ -76,24 +85,135 @@ export class UserCrudController
 		res: Response,
 		next: NextFunction
 	): Promise<void> {
-		const { email, id, isAdmin } = verify(
-			req.token,
-			process.env.ACCESS_TOKEN_SECRET!
-		) as IUserPayload;
-		const result = await this.userService.updateUser(
-			{
+		const accessTokenData = this.jwtService.verifyAccessToken(
+			req.cookies.accessToken
+		);
+		if (!accessTokenData) {
+			const refreshTokenData = this.jwtService.verifyRefreshToken(
+				req.cookies.refreshToken
+			);
+			if (!refreshTokenData) {
+				res.status(401).json({
+					message: "Unauthorized. Log in, please",
+				});
+				return;
+			}
+			const newAccessToken = await this.jwtService.signAccessToken(
+				refreshTokenData.email,
+				Number(refreshTokenData.id),
+				refreshTokenData.isAdmin
+			);
+			const result = await this.userService.updateUser({
 				firstName: req.body.firstName,
 				lastName: req.body.lastName,
 				email: req.body.email,
 				password: req.body.password,
-			},
-			req.body.updateId,
-			Number(id) //can pass here user id and create another method with custom id for admin
-		);
-		if (result) {
-			res.json({ message: "User updated" });
+				isAdmin: refreshTokenData.isAdmin,
+				updateId: Number(refreshTokenData.id),
+			});
+			if (result) {
+				res.cookie("accessToken", newAccessToken);
+				res.json({ message: "User updated" });
+			} else {
+				res.cookie("accessToken", newAccessToken);
+				res.status(400).json({ message: "User not updated" });
+			}
 		} else {
-			res.status(400).json({ message: "User not updated" });
+			const result = await this.userService.updateUser({
+				firstName: req.body.firstName,
+				lastName: req.body.lastName,
+				email: req.body.email,
+				password: req.body.password,
+				isAdmin: accessTokenData.isAdmin,
+				updateId: Number(accessTokenData.id),
+			});
+			if (result) {
+				res.json({ message: "User updated" });
+			} else {
+				res.status(400).json({ message: "User not updated" });
+			}
+		}
+	}
+
+	async updateAny(
+		req: Request,
+		res: Response,
+		next: NextFunction
+	): Promise<void> {
+		const accessTokenData = this.jwtService.verifyAccessToken(
+			req.cookies.accessToken
+		);
+		if (!accessTokenData) {
+			const refreshTokenData = this.jwtService.verifyRefreshToken(
+				req.cookies.refreshToken
+			);
+			if (!refreshTokenData) {
+				res.status(401).json({
+					message: "Unauthorized. Log in, please",
+				});
+				return;
+			}
+			const newAccessToken = await this.jwtService.signAccessToken(
+				refreshTokenData.email,
+				Number(refreshTokenData.id),
+				refreshTokenData.isAdmin
+			);
+			const result = await this.userService.updateUser(req.body);
+			if (result) {
+				res.cookie("accessToken", newAccessToken);
+				res.json({ message: "User updated" });
+			} else {
+				res.cookie("accessToken", newAccessToken);
+				res.status(400).json({ message: "User not updated" });
+			}
+		} else {
+			const result = await this.userService.updateUser(req.body);
+			if (result) {
+				res.json({ message: "User updated" });
+			} else {
+				res.status(400).json({ message: "User not updated" });
+			}
+		}
+	}
+
+	async deleteAny(
+		req: Request,
+		res: Response,
+		next: NextFunction
+	): Promise<void> {
+		const accessTokenData = this.jwtService.verifyAccessToken(
+			req.cookies.accessToken
+		);
+		if (!accessTokenData) {
+			const refreshTokenData = this.jwtService.verifyRefreshToken(
+				req.cookies.refreshToken
+			);
+			if (!refreshTokenData) {
+				res.status(401).json({
+					message: "Unauthorized. Log in, please",
+				});
+				return;
+			}
+			const newAccessToken = await this.jwtService.signAccessToken(
+				refreshTokenData.email,
+				Number(refreshTokenData.id),
+				refreshTokenData.isAdmin
+			);
+			const result = await this.userService.deleteUser(req.body.deleteId);
+			if (result) {
+				res.cookie("accessToken", newAccessToken);
+				res.json({ message: "User deleted" });
+			} else {
+				res.cookie("accessToken", newAccessToken);
+				res.status(400).json({ message: "User not deleted" });
+			}
+		} else {
+			const result = await this.userService.deleteUser(req.body.deleteId);
+			if (result) {
+				res.json({ message: "User deleted" });
+			} else {
+				res.status(400).json({ message: "User not deleted" });
+			}
 		}
 	}
 
@@ -102,19 +222,43 @@ export class UserCrudController
 		res: Response,
 		next: NextFunction
 	): Promise<void> {
-		const { email, id, isAdmin } = verify(
-			req.cookies.accessToken,
-			process.env.ACCESS_TOKEN_SECRET!
-		) as IUserPayload;
-		const result = await this.userService.deleteUser(
-			Number(id),
-			req.body.deleteId,
-			isAdmin
+		const accessTokenData = this.jwtService.verifyAccessToken(
+			req.cookies.accessToken
 		);
-		if (result) {
-			res.json({ message: "User deleted" });
+		if (!accessTokenData) {
+			const refreshTokenData = this.jwtService.verifyRefreshToken(
+				req.cookies.refreshToken
+			);
+			if (!refreshTokenData) {
+				res.status(401).json({
+					message: "Unauthorized. Log in, please",
+				});
+				return;
+			}
+			const newAccessToken = await this.jwtService.signAccessToken(
+				refreshTokenData.email,
+				Number(refreshTokenData.id),
+				refreshTokenData.isAdmin
+			);
+			const result = await this.userService.deleteUser(
+				Number(refreshTokenData.id)
+			);
+			if (result) {
+				res.cookie("accessToken", newAccessToken);
+				res.json({ message: "User deleted" });
+			} else {
+				res.cookie("accessToken", newAccessToken);
+				res.status(400).json({ message: "User not deleted" });
+			}
 		} else {
-			res.status(400).json({ message: "User not deleted" });
+			const result = await this.userService.deleteUser(
+				Number(accessTokenData.id)
+			);
+			if (result) {
+				res.json({ message: "User deleted" });
+			} else {
+				res.status(400).json({ message: "User not deleted" });
+			}
 		}
 	}
 }
